moved pkg/discovery to internal/discovery
