+++ internal/discovery/notifier_test.go
