+++ internal/discovery/package_test.go
